+++ Core/src/com/mfino/domain/CommodityTransfer.java
@@ -97,7 +97,7 @@ public class CommodityTransfer extends Base implements java.io.Serializable {
 	private String sourcecardpan;
 	private String sourcemessage;
 	private String destmdn;
-	private BigDecimal destmdnid;
+	private Long destmdnid;
 	private BigDecimal destsubscriberid;
 	private String destsubscribername;
 	private Long destpocketallowance;
@@ -455,11 +455,11 @@ public class CommodityTransfer extends Base implements java.io.Serializable {
 	}
 
 	@Column(name = "DESTMDNID", scale = 0)
-	public BigDecimal getDestmdnid() {
+	public Long getDestmdnid() {
 		return this.destmdnid;
 	}
 
-	public void setDestmdnid(BigDecimal destmdnid) {
+	public void setDestmdnid(Long destmdnid) {
 		this.destmdnid = destmdnid;
 	}
